requirements.txt
beautifulsoup4
Pyrogram~=1.2.9
python-dotenv~=0.18.0
requests
selenium
youtube_dl